requirements.txt
imageio==2.36.1
imageio-ffmpeg==0.5.1
numpy==2.2.1
pillow==11.1.0
psutil==6.1.1
